requirements.txt
django==4.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt
mysqlclient

